mtd: spi-nor: fix incorrect len return
in nor_write, we need to say back to upper layers about how much
data we transferred. there is a bug in this logic while updating the
number of bytes written. this patch fixes this issue. the existing logic
is not checking for last page and from there onwards we are running into
issues while updating the written bytes.

Signed-off-by: Naga Sureshkumar Relli <nagasure@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>