iio: ad9517: Fix potential out-of-bounds access in firmware parser
The firmware parser uses string handling functions like strchr() and
scanf(). These functions expect that the data is zero terminated, but the
firmware data buffer is not zero terminated. This means those functions
might cause an out-of-bounds access on the buffer.

Fix this by adding a zero byte at the end of the firmware buffer.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>